+++ integrations/sonarqube/CHANGELOG.md
@@ -7,12 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
-## 0.1.168 (2025-04-25)
+## 0.1.169 (2025-04-15)
 
 
 ### Bug Fixes
 
 - Fixed bug where the integration was trying to fetch more than 10,000 issues from SonarQube API
+
+
+## 0.1.168 (2025-04-25)
+
+
+### Bug Fixes
+
+- Fixed handling of `base_url` and `app_host` to strip trailing slashes.
+- Corrected `app_host=None` case to set `webhook_invoke_url` to an empty string.
 
 
 ## 0.1.167 (2025-04-15)